+++ lib/eve_online/esi/models/jump_clone.rb
@@ -6,7 +6,6 @@ module EveOnline
       class JumpClone < Base
         def as_json
           {
-            implants: implants,
             jump_clone_id: jump_clone_id,
             location_id: location_id,
             location_type: location_type,
@@ -14,6 +13,7 @@ module EveOnline
           }
         end
 
+        # TODO: rename to implant_ids
         def implants
           options['implants']
         end